:truck: Refs open-formulieren/open-forms#928 - Move summary util functions into their own file
