+++ src/components/FormStepSummary/index.js
@@ -9,7 +9,7 @@ import { Table, TableRow, TableHead, TableCell } from 'components/Table';
 import { Toolbar, ToolbarList } from 'components/Toolbar';
 
 import {getComponentLabel, getComponentValue,
-  iterComponentKeyValues, getComponentModifiers, displayValue} from 'utils';
+  iterComponentKeyValues, getComponentModifiers, displayValue} from 'components/FormStepSummary/utils';
 
 
 const FormStepSummary = ({stepData, editStepUrl, editStepText}) => {
